Fix excluding files for submodules
The previous exclude file path contained a hardcoded git repository path: `.git`.
It's correct for simple repositories but it's invalid for the submodules.

The output of the `git rev-parse --git-dir` should be used as the git directory path instead.
The command outputs a proper path for both simple repositories and submodules, e.g.:
`<parent-repository-path>/.git/modules/<submodule-name>`.